backend/requirements.txt
# FastAPI and web
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis and Celery
redis==5.0.1
celery==5.3.6

# Pydantic
pydantic==2.6.1
pydantic-settings==2.1.0

# LLM Providers
openai==1.12.0
anthropic==0.18.1
groq==0.4.2

# LangGraph and LangChain
langgraph==0.0.26
langchain==0.1.9
langchain-openai==0.0.6
langchain-anthropic==0.1.1
langchain-community==0.0.24

# Code parsing
tree-sitter==0.21.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.0
tree-sitter-typescript==0.21.0

# Graph and embeddings
networkx==3.2.1
chromadb==0.4.22
sentence-transformers==2.3.1

# Git operations
gitpython==3.1.42

# Utilities
httpx==0.26.0
python-dotenv==1.0.1
tenacity==8.2.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0
